Move packageInfo to models
