+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Helpers/PackageInfo/GoPackageInfoHelper.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 using System.Text.RegularExpressions;
+using Azure.Sdk.Tools.Cli.Models;
 
 namespace Azure.Sdk.Tools.Cli.Helpers;
