+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Helpers/PackageInfo/IPackageInfoHelper.cs
@@ -1,5 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
+using Azure.Sdk.Tools.Cli.Models;
+
 namespace Azure.Sdk.Tools.Cli.Helpers;
 
 /// <summary>
