+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Helpers/PackageInfo/JavaScriptPackageInfoHelper.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 using System.Text.Json;
+using Azure.Sdk.Tools.Cli.Models;
 
 namespace Azure.Sdk.Tools.Cli.Helpers;
