+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Models/PackageInfo.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-using Azure.Sdk.Tools.Cli.Models;
-
-namespace Azure.Sdk.Tools.Cli.Helpers;
+namespace Azure.Sdk.Tools.Cli.Models;
 
 /// <summary>
 /// Plain data model representing inferred information about an Azure SDK package.
